// src/app/model/fieldLists.ts
export interface FieldLists {
    countryFields: string[];
    gblProgramFields: string[];
    partnerNameFields: string[];
    partnerLocationFields: string[];
    modelOfCollaborationTypeFields: string[];

    countryFieldsCounts: number[];
    gblProgramFieldsCounts: number[];
    partnerNameFieldsCounts: number[]; 
    partnerLocationFieldsCounts: number[];
    modelOfCollaborationTypeFieldsCounts: number[];
}